Move game reset functionality into main App
