+++ src/components/Outcome.js
@@ -54,43 +54,10 @@ const Answer = styled.span`
   text-align: center;
 `
 
-export const Outcome = ({ 
-  outcome, setOutcome, 
-  answer, setAnswer, getAnswer,
-  feedback, setFeedback, 
-  guesses, setGuesses }) => {
+export const Outcome = ({ outcome, setOutcome, answer, startNewGame }) => {
 
   const closeModal = () => {
     setOutcome('');
-  }
-
-  const startNewGame = () => {
-    setOutcome('');
-    setFeedback({
-      0: Array.from({ length: 2 }).fill(''),
-      1: Array.from({ length: 2 }).fill(''),
-      2: Array.from({ length: 2 }).fill(''),
-      3: Array.from({ length: 2 }).fill(''),
-      4: Array.from({ length: 2 }).fill(''),
-      5: Array.from({ length: 2 }).fill(''),
-      6: Array.from({ length: 2 }).fill(''),
-      7: Array.from({ length: 2 }).fill(''),
-      8: Array.from({ length: 2 }).fill(''),
-      9: Array.from({ length: 2 }).fill(''),
-    });
-    setGuesses({    
-      0: Array.from({ length: 4 }).fill(''),
-      1: Array.from({ length: 4 }).fill(''),
-      2: Array.from({ length: 4 }).fill(''),
-      3: Array.from({ length: 4 }).fill(''),
-      4: Array.from({ length: 4 }).fill(''),
-      5: Array.from({ length: 4 }).fill(''),
-      6: Array.from({ length: 4 }).fill(''),
-      7: Array.from({ length: 4 }).fill(''),
-      8: Array.from({ length: 4 }).fill(''),
-      9: Array.from({ length: 4 }).fill(''),
-    });
-    setAnswer(getAnswer());
   }
 
   return (
