+++ src/util/GameDefaults.js
@@ -1,0 +1,25 @@
+export const initBoard = {
+  0: Array.from({ length: 4 }).fill(''),
+  1: Array.from({ length: 4 }).fill(''),
+  2: Array.from({ length: 4 }).fill(''),
+  3: Array.from({ length: 4 }).fill(''),
+  4: Array.from({ length: 4 }).fill(''),
+  5: Array.from({ length: 4 }).fill(''),
+  6: Array.from({ length: 4 }).fill(''),
+  7: Array.from({ length: 4 }).fill(''),
+  8: Array.from({ length: 4 }).fill(''),
+  9: Array.from({ length: 4 }).fill(''),
+}
+
+export const initFeedback = {
+  0: Array.from({ length: 2 }).fill(''),
+  1: Array.from({ length: 2 }).fill(''),
+  2: Array.from({ length: 2 }).fill(''),
+  3: Array.from({ length: 2 }).fill(''),
+  4: Array.from({ length: 2 }).fill(''),
+  5: Array.from({ length: 2 }).fill(''),
+  6: Array.from({ length: 2 }).fill(''),
+  7: Array.from({ length: 2 }).fill(''),
+  8: Array.from({ length: 2 }).fill(''),
+  9: Array.from({ length: 2 }).fill(''),
+}
